+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Pending
 
+* Restore external end screen usage when available [#15537](https://github.com/Applifier/unity-ads-webview/pull/15537)
+* Add support for data url in external end screen [#15524](https://github.com/Applifier/unity-ads-webview/pull/15524)
+* Add getDataUrl to XHRequest.ts [#15523](https://github.com/Applifier/unity-ads-webview/pull/15523)
+
+# Thu May 28 23:11:17 UTC 2020
+
+* [Load V5] Add load v5 support flag to ad request body [#15528](https://github.com/Applifier/unity-ads-webview/pull/15528)
 * [Load V5] Add Zynga games to load v5 [#15531](https://github.com/Applifier/unity-ads-webview/pull/15531)
 * [Load V5] Always invalidate programatic campaigns if Ad Unit is in use. [#15527](https://github.com/Applifier/unity-ads-webview/pull/15527)
 * [AUI] Refactor renderColorTheme for reusability in the ColorBlur experiment [#15365](https://github.com/Applifier/unity-ads-webview/pull/15365)